+++ deepinv/datasets/utils.py
@@ -6,6 +6,8 @@ import zipfile
 import requests
 from tqdm.auto import tqdm
 
+from torch.utils.data import Dataset
+from torch import randn
 
 def check_path_is_a_folder(folder_path: str) -> bool:
     # Check if `folder_path` is pointing to a directory
@@ -27,7 +29,7 @@ def calculate_md5(fpath: str, chunk_size: int = 1024 * 1024) -> str:
     return md5.hexdigest()
 
 
-def calculate_md5_for_folder(folder_path: str):
+def calculate_md5_for_folder(folder_path: str) -> str:
     """Compute the hash of all files in a folder then compute the hash of the folder.
 
     Folder will be considered as empty if it is not strictly containing files.
@@ -40,7 +42,7 @@ def calculate_md5_for_folder(folder_path: str):
     return md5_folder.hexdigest()
 
 
-def download_zipfile(url, save_path):
+def download_zipfile(url: str, save_path: str) -> None:
     """Download zipfile from the Internet."""
     # Ensure the directory containing `save_path`` exists
     os.makedirs(os.path.dirname(save_path), exist_ok=True)
@@ -58,7 +60,7 @@ def download_zipfile(url, save_path):
     del response
 
 
-def extract_zipfile(file_path, extract_dir):
+def extract_zipfile(file_path, extract_dir) -> None:
     """Extract a local zipfile."""
     # Open the zip file
     with zipfile.ZipFile(file_path, "r") as zip_ref:
@@ -67,3 +69,22 @@ def extract_zipfile(file_path, extract_dir):
         # thus the progres bar will not move linearly with time
         for file_to_be_extracted in tqdm(zip_ref.infolist(), desc="Extracting"):
             zip_ref.extract(file_to_be_extracted, extract_dir)
+
+class PlaceholderDataset(Dataset):
+    """
+    A placeholder dataset for test purposes.
+
+    Produces image pairs x,y that are random tensor of shape specified.
+    
+    :param int n: number of samples in dataset, defaults to 1
+    :param tuple shape: image shape, (channel, height, width), defaults to (1, 64, 64)
+    """
+    def __init__(self, n=1, shape=(1, 64, 64)):
+        self.n = n
+        self.shape = shape
+    
+    def __len__(self):
+        return self.n
+    
+    def __getitem__(self, index):
+        return randn(self.shape), randn(self.shape)