+++ deepinv/physics/forward.py
@@ -261,7 +261,9 @@ class Physics(torch.nn.Module):  # parent class for forward models
                 "update_parameters method not implemented for this physics operator"
             )
 
-        if self.noise_model is not None:
+        # if self.noise_model is not None:
+        # check if noise model has a method named update_parameters
+        if hasattr(self.noise_model, "update_parameters"):
             self.noise_model.update_parameters(**kwargs)
 
 
@@ -678,7 +680,8 @@ class DecomposablePhysics(LinearPhysics):
         :return: (torch.Tensor) output tensor
 
         """
-        self.update_parameters(mask=mask)
+
+        self.update_parameters(mask=mask, **kwargs)
 
         return self.U(self.mask * self.V_adjoint(x))
 
@@ -694,7 +697,7 @@ class DecomposablePhysics(LinearPhysics):
         :return: (torch.Tensor) output tensor
         """
 
-        self.update_parameters(mask=mask)
+        self.update_parameters(mask=mask, **kwargs)
 
         if isinstance(self.mask, float):
             mask = self.mask
@@ -734,7 +737,7 @@ class DecomposablePhysics(LinearPhysics):
         x = self.V(self.V_adjoint(b) / scaling)
         return x
 
-    def A_dagger(self, y, **kwargs):
+    def A_dagger(self, y, mask=None, **kwargs):
         r"""
         Computes :math:`A^{\dagger}y = x` in an efficient manner leveraging the singular vector decomposition.
 
@@ -742,6 +745,9 @@ class DecomposablePhysics(LinearPhysics):
         :return: (torch.Tensor) The reconstructed image :math:`x`.
 
         """
+
+        # TODO should this happen here or at the end of A_dagger?
+        self.update_parameters(mask=mask, **kwargs)
 
         # avoid division by singular value = 0