+++ CHANGELOG.rst
@@ -8,8 +8,10 @@ This change log is for the `main` branch. It contains changes for each release, 
 Current
 ----------------
 
+
 New Features
 ^^^^^^^^^^^^
+- Add support for adversarial losses and models (GANs) (:gh:`183` by `Andrew Wang`_)
 - Base transform class for transform arithmetic (:gh:`240` by `Andrew Wang`_) - 26/06/2024.
 - Plot video/animation functionality (:gh:`245` by `Andrew Wang`_)
 - Added update_parameters for parameter-dependent physics (:gh:`241` by Julian Tachella) - 11/06/2024
@@ -21,7 +23,7 @@ New Features
 Fixed
 ^^^^^
 - Fixed averaging loss over epochs Trainer (:gh:`241` by Julian Tachella) - 11/06/2024
-
+- Fixed Trainer save_path timestamp problem on Windows (:gh:`245` by `Andrew Wang`_)
 
 Changed
 ^^^^^^^
@@ -63,6 +65,7 @@ New Features
 - Added Restormer model (:gh:`185` by Antoine Regnier and Maxime SONG) - 18/04/2024
 - Added DIV2K built-in dataset (:gh:`203` by Maxime SONG) - 03/05/2024
 - Added Urban100 built-in dataset (:gh:`237` by Maxime SONG) - 07/06/2024
+- Added Set14 / CBSD68 / fastMRI buit-in datasets (:gh:`248` :gh:`249` :gh:`229` by Maxime SONG) - 25/06/2024 
 
 Fixed
 ^^^^^